@@ -3,10 +3,9 @@ import { NbCalendarRange, NbDateService } from '@nebular/theme';
 import { DayCellComponent } from './day-cell/day-cell.component';
 
 @Component({
-  selector: 'ngx-calendar',
-  templateUrl: 'calendar.component.html',
-  styleUrls: ['calendar.component.scss'],
-  entryComponents: [DayCellComponent],
+    selector: 'ngx-calendar',
+    templateUrl: 'calendar.component.html',
+    styleUrls: ['calendar.component.scss'],
 })
 export class CalendarComponent {
 
