@@ -77,18 +77,15 @@ const SERVICES = [
 ];
 
 @NgModule({
-  imports: [
-    ...MODULES,
-  ],
-  declarations: [
-    ...COMPONENTS,
-  ],
-  providers: [
-    ...SERVICES,
-  ],
-  entryComponents: [
-    ...ENTRY_COMPONENTS,
-  ],
+    imports: [
+        ...MODULES,
+    ],
+    declarations: [
+        ...COMPONENTS,
+    ],
+    providers: [
+        ...SERVICES,
+    ],
 })
 export class ModalOverlaysModule {
 }
